requirements.txt
numpy
opencv-python
PyYAML
scikit-learn
ultralytics